Add usecurrencyField hook
