+++ src/components/Other/index.ts
@@ -1,0 +1,1 @@
+export { CurrencyField } from './CurrencyField/CurrencyField';
